Renamed Components folder as Pages and Added New SplashScreen and Auth Screen
