--- .env
@@ -1,3 +1,0 @@
-NEXT_PUBLIC_APP_URL=http://localhost:1420
-
-DATABASE_URL="file:./dev.db"
