+++ .env-example
@@ -1,0 +1,4 @@
+NEXT_PUBLIC_APP_URL=http://localhost:1420
+
+DATABASE_URL="file:./dev.db"
+AUTH_SECRET="mysecret"
