+++ auth.config.ts
@@ -1,0 +1,7 @@
+import type { NextAuthConfig } from 'next-auth';
+
+export const authConfig = {
+    pages: {
+        signIn: '/login',
+    },
+} as NextAuthConfig;
