+++ package.json
@@ -31,6 +31,7 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@heroicons/react": "^2.1.1",
     "@hookform/resolvers": "^3.3.4",
+    "@lottiefiles/react-lottie-player": "^3.5.3",
     "@prisma/extension-accelerate": "^0.6.3",
     "@radix-ui/react-accessible-icon": "^1.0.3",
     "@radix-ui/react-accordion": "^1.1.2",
@@ -63,7 +64,7 @@
     "@radix-ui/react-tooltip": "^1.0.7",
     "@radix-ui/themes": "^2.0.3",
     "@tailwindcss/line-clamp": "^0.4.4",
-    "@tanstack/react-table": "^8.11.7",
+    "@tanstack/react-table": "^8.11.8",
     "@tauri-apps/api": "2.0.0-alpha.5",
     "@tauri-apps/plugin-app": "2.0.0-alpha.0",
     "@tauri-apps/plugin-os": "2.0.0-alpha.0",
@@ -85,6 +86,7 @@
     "lucide-react": "^0.315.0",
     "nanoid": "^5.0.4",
     "next": "^14.1.0",
+    "next-auth": "^4.24.5",
     "next-themes": "^0.2.1",
     "next-transpile-modules": "^9.0.0",
     "node-fetch": "^3.3.2",
@@ -98,7 +100,7 @@
     "react-icons": "^4.4.0",
     "react-markdown": "^9.0.1",
     "react-modal": "^3.16.1",
-    "react-router-dom": "^6.22.0",
+    "react-router-dom": "^6.21.3",
     "react-scripts": "5.0.1",
     "react-spring": "^9.7.3",
     "react-tabs": "^6.0.2",
@@ -121,6 +123,7 @@
     "@types/node": "^20.11.17",
     "@types/react": "^18.2.48",
     "@types/react-dom": "^18.2.19",
+    "@types/react-lottie": "^1.2.10",
     "@types/react-modal": "^3.16.3",
     "autoprefixer": "^10.4.17",
     "clsx": "^2.1.0",
