+++ prettier.config.js
@@ -15,8 +15,8 @@ module.exports = {
     "^@/config/(.*)$",
     "^@/lib/(.*)$",
     "^@/hooks/(.*)$",
-    "^@/components/ui/(.*)$",
-    "^@/components/(.*)$",
+    "^@/pages/ui/(.*)$",
+    "^@/pages/(.*)$",
     "^@/styles/(.*)$",
     "^@/app/(.*)$",
     "",
