+++ public/lottie/loading2.json
@@ -1,0 +1,1 @@
+{"v":"5.0.1","fr":10,"ip":0,"op":20,"w":800,"h":600,"ddd":0,"assets":[],"layers":[{"ind":3,"nm":"Layer 3","ks":{"p":{"a":1,"k":[{"t":0,"s":[456.419,423.252,0],"to":[-161.825,-53.536,0],"ti":[-62.867,94.219,0],"i":{"x":[0.15],"y":[0.86]},"o":{"x":[0.785],"y":[0.135]},"e":[371.703,167.513,0]},{"t":10,"s":[371.703,167.513,0],"to":[0,0,0],"ti":[0,0,0],"i":{"x":[0.15],"y":[0.86]},"o":{"x":[0.785],"y":[0.135]},"e":[456.419,423.252,0]},{"t":20,"s":[456.419,423.252,0]}]},"a":{"a":0,"k":[-24.5,-24.5,0]},"s":{"a":0,"k":[100,100,100]},"r":{"a":0,"k":-43.065},"o":{"a":0,"k":100}},"ao":0,"ip":0,"op":20,"st":0,"bm":3,"sr":1,"ty":4,"shapes":[{"ty":"gr","it":[{"d":1,"ty":"el","s":{"a":0,"k":[210.919,210.919]},"p":{"a":0,"k":[0,0]}},{"ty":"fl","c":{"a":0,"k":[0,0.698,1,1]},"hd":false,"o":{"a":0,"k":100},"r":1},{"ty":"tr","p":{"a":0,"k":[0,0]},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"r":{"a":0,"k":0},"o":{"a":0,"k":100},"sk":{"a":0,"k":0},"sa":{"a":0,"k":0}}],"nm":"Object","hd":false}]},{"ind":2,"nm":"Layer 2","ks":{"p":{"a":1,"k":[{"t":0,"s":[361.777,165.918,0],"to":[0,0,0],"ti":[0,0,0],"i":{"x":[0.15],"y":[0.86]},"o":{"x":[0.785],"y":[0.135]},"e":[214.784,406.838,0]},{"t":10,"s":[214.784,406.838,0],"to":[-83.035,-34.503,0],"ti":[0,0,0],"i":{"x":[0.15],"y":[0.86]},"o":{"x":[0.785],"y":[0.135]},"e":[361.777,165.918,0]},{"t":20,"s":[361.777,165.918,0]}]},"a":{"a":0,"k":[-66.5,-66.5,0]},"s":{"a":0,"k":[100,100,100]},"r":{"a":0,"k":-43.065},"o":{"a":0,"k":100}},"ao":0,"ip":0,"op":20,"st":0,"bm":3,"sr":1,"ty":4,"shapes":[{"ty":"gr","it":[{"d":1,"ty":"el","s":{"a":0,"k":[236,236]},"p":{"a":0,"k":[0,0]}},{"ty":"fl","c":{"a":0,"k":[0.012,0,0.38,1]},"hd":false,"o":{"a":0,"k":100},"r":1},{"ty":"tr","p":{"a":0,"k":[0,0]},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"r":{"a":0,"k":0},"o":{"a":0,"k":100},"sk":{"a":0,"k":0},"sa":{"a":0,"k":0}}],"nm":"Object","hd":false}]},{"ind":1,"nm":"Layer 1","ks":{"p":{"a":1,"k":[{"t":0,"s":[263.776,342.564,0],"to":[0,0,0],"ti":[0,0,0],"i":{"x":[0.15],"y":[0.86]},"o":{"x":[0.785],"y":[0.135]},"e":[545.716,355.446,0]},{"t":10,"s":[545.716,355.446,0],"to":[-7.345,74.699,0],"ti":[0,0,0],"i":{"x":[0.15],"y":[0.86]},"o":{"x":[0.785],"y":[0.135]},"e":[263.776,342.564,0]},{"t":20,"s":[263.776,342.564,0]}]},"a":{"a":0,"k":[0,0,0]},"s":{"a":0,"k":[100,100,100]},"r":{"a":0,"k":-43.065},"o":{"a":0,"k":100}},"ao":0,"ip":0,"op":20,"st":0,"bm":3,"sr":1,"ty":4,"shapes":[{"ty":"gr","it":[{"d":1,"ty":"el","s":{"a":0,"k":[150.135,150.135]},"p":{"a":0,"k":[0,0]}},{"ty":"fl","c":{"a":0,"k":[0.298,0.255,0.714,1]},"hd":false,"o":{"a":0,"k":100},"r":1},{"ty":"tr","p":{"a":0,"k":[0,0]},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"r":{"a":0,"k":0},"o":{"a":0,"k":100},"sk":{"a":0,"k":0},"sa":{"a":0,"k":0}}],"nm":"Object","hd":false}]}],"markers":[]}
