+++ src/app/auth/layout.tsx
@@ -32,6 +32,17 @@ export default function MyApp({ children }: ExamplesLayoutProps) {
       <body className="overflow-clip bg-transparent font-sans antialiased scrollbar-none">
       <Theme>
 
+          <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
+              <div className="h-screen overflow-clip">
+                  <Menu menuItems={false}/>
+                  <div className="flex h-full w-full bg-background-100 dark:bg-background-900">
+                      {/* Navbar & Main Content */}
+                      <div className="h-full w-full font-dm dark:bg-navy-900">
+                          {/* Main Content */}
+                          <main
+                              className={`mx-2.5  flex-none transition-all dark:bg-navy-900 
+              md:pr-2 xl:ml-[323px] `}
+                          >
 
 
 
@@ -40,7 +51,11 @@ export default function MyApp({ children }: ExamplesLayoutProps) {
                                       {children}
                                   </div>
 
-
+                          </main>
+                      </div>
+                  </div>
+              </div>
+          </ThemeProvider>
 
       </Theme>
       <StyleSwitcher/>
