+++ src/app/auth/page.tsx
@@ -1,86 +1,69 @@
 import { Metadata } from "next"
 import Image from "next/image"
 import Link from "next/link"
-
 import { cn } from "@/lib/utils"
 import { buttonVariants } from "@/components/ui/button"
 import { UserAuthForm } from "@/app/auth/components/user-auth-form"
 
 export const metadata: Metadata = {
   title: "Authentication",
-  description: "Authentication forms built using the components.",
+  description: "Authentication forms built using the pages.",
 }
 
 export default function AuthenticationPage() {
   return (
-    <>
-      <div className="container relative  h-[800px] flex-col items-center justify-center md:grid lg:max-w-none lg:grid-cols-2 lg:px-0">
-        <Link
-          href="/auth"
-          className={cn(
-            buttonVariants({ variant: "ghost" }),
-            "absolute right-4 top-4 md:right-8 md:top-8"
-          )}
-        >
-          Login
-        </Link>
-        <div className="relative  h-full flex-col bg-muted p-10 text-white dark:border-r lg:flex">
-          <div className="absolute inset-0 bg-zinc-900" />
-          <div className="relative z-20 flex items-center text-lg font-medium">
+      <>
+        <div className="container relative flex items-center justify-center lg:max-w-none lg:px-0 min-h-screen">
+          <div className="lg:w-1/2 md:w-full lg:fixed lg:left-0 lg:top-10 lg:h-full lg:bg-background border-t border-primary/10 border lg:text-primary lg:flex lg:flex-col lg:justify-center lg:items-center lg:py-16 lg:px-8">
             <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              className="mr-2 h-6 w-6"
+                xmlns="http://www.w3.org/2000/svg"
+                viewBox="0 0 24 24"
+                fill="none"
+                stroke="currentColor"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                className="mb-4 h-12 w-12 text-primary"
             >
               <path d="M15 6v12a3 3 0 1 0 3-3H6a3 3 0 1 0 3 3V6a3 3 0 1 0-3 3h12a3 3 0 1 0-3-3" />
             </svg>
-            Maaagic AI
-          </div>
-          <div className="relative z-20 mt-auto">
-            <blockquote className="space-y-2">
-              <p className="text-lg">
-                &ldquo;This Open Source makes you build your AI Application in rapid Speed&rdquo;
-              </p>
-              <footer className="text-sm">anonymous</footer>
+            <h1 className="text-4xl font-bold mb-4 text-center lg:text-left">Maaagic AI</h1>
+            <blockquote className="text-lg text-center lg:text-left">
+              “This Open Source makes you integrate applications in rapid Speed”
+              <footer className="text-sm">- Muhammad Fiaz</footer>
             </blockquote>
           </div>
-        </div>
-        <div className="lg:p-8">
-          <div className="mx-auto flex w-full flex-col justify-center space-y-6 sm:w-[350px]">
-            <div className="flex flex-col space-y-2 text-center">
-              <h1 className="text-2xl font-semibold tracking-tight">
-                Create an account
-              </h1>
-              <p className="text-sm text-muted-foreground">
-                Enter your email below to create your account
+          <div className="lg:w-1/2 md:w-full lg:ml-auto lg:px-8">
+            <div className="mx-auto flex flex-col justify-center space-y-6 sm:max-w-[350px] lg:w-full lg:max-w-[400px]">
+              <div className="flex flex-col space-y-2 text-center lg:text-left">
+                <h1 className="text-2xl font-semibold tracking-tight">
+                  Sign in to your account
+                </h1>
+                <p className="text-sm text-muted-foreground">
+                  Enter your email below to sign in your account
+                </p>
+              </div>
+              <UserAuthForm />
+              <p className="px-8 text-center text-sm text-muted-foreground">
+                By clicking continue, you agree to our{" "}
+                <Link
+                    href="/terms"
+                    className="underline underline-offset-4 hover:text-primary"
+                >
+                  Terms of Service
+                </Link>{" "}
+                and{" "}
+                <Link
+                    href="/privacy"
+                    className="underline underline-offset-4 hover:text-primary"
+                >
+                  Privacy Policy
+                </Link>
+                .
               </p>
             </div>
-            <UserAuthForm />
-            <p className="px-8 text-center text-sm text-muted-foreground">
-              By clicking continue, you agree to our{" "}
-              <Link
-                href="/terms"
-                className="underline underline-offset-4 hover:text-primary"
-              >
-                Terms of Service
-              </Link>{" "}
-              and{" "}
-              <Link
-                href="/privacy"
-                className="underline underline-offset-4 hover:text-primary"
-              >
-                Privacy Policy
-              </Link>
-              .
-            </p>
           </div>
         </div>
-      </div>
-    </>
+      </>
   )
 }
