--- src/app/components/layout.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-interface ExamplesLayoutProps {
-  children: React.ReactNode
-}
-
-export default function ExamplesLayout({ children }: ExamplesLayoutProps) {
-  return <div>{children}
-  </div>
-}
