+++ src/app/pages/page.tsx
@@ -1,0 +1,10 @@
+import MaaagicAI from "@/app/pages/Maaagic_AI/page";
+
+export default function App() {
+
+  return (
+      <div>
+<MaaagicAI/>
+      </div>
+  )
+}
