+++ src/app/pages/settings/account/page.tsx
@@ -1,5 +1,5 @@
 import { Separator } from "@/components/ui/separator"
-import { AccountForm } from "@/app/components/settings/account/account-form"
+import { AccountForm } from "@/app/pages/settings/account/account-form"
 
 export default function SettingsAccountPage() {
   return (
