+++ src/app/pages/settings/display/display-form.tsx
