+++ src/app/pages/settings/page.tsx
@@ -1,5 +1,5 @@
 import { Separator } from "@/components/ui/separator"
-import { ProfileForm } from "@/app/components/settings/components/profile-form"
+import { ProfileForm } from "@/app/pages/settings/components/profile-form"
 
 export default function SettingsProfilePage() {
   return (
